+++ src/main/java/com/archers_expansion/client/PinDownParticles.java
@@ -12,7 +12,7 @@ public class PinDownParticles implements CustomParticleStatusEffect.Spawner{
         this.particles = new ParticleBatch(
                 "spell_engine:weakness_smoke",
                 ParticleBatch.Shape.CIRCLE, ParticleBatch.Origin.FEET,
-                null, particleCount, 0.1F, 0.1F, 360);
+                null, particleCount, 0.1F, 0.2F, 360);
     }
 
     @Override