+++ src/main/java/com/archers_expansion/effect/PointBlankShot.java
@@ -1,12 +1,12 @@
 package com.archers_expansion.effect;
 
+import net.minecraft.enchantment.EnchantmentHelper;
+import net.minecraft.enchantment.Enchantments;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.attribute.AttributeContainer;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectCategory;
 import net.minecraft.util.Identifier;
-import net.minecraft.util.math.Vec3d;
-import net.spell_engine.api.spell.SpellInfo;
 
 import static com.archers_expansion.ArchersExpansionMod.MOD_ID;
 import static net.spell_engine.internals.SpellRegistry.getSpell;
@@ -20,10 +20,6 @@ public class PointBlankShot extends StatusEffect {
     public void onApplied(LivingEntity entity, AttributeContainer attributes, int amplifier) {
         if (!entity.getWorld().isClient) {
             LivingEntity attacker = entity.getLastAttacker();
-            entity.velocityDirty = true;
-            entity.velocityModified = true;
-            Vec3d rotationVector = entity.getRotationVector();
-            Vec3d velocity = entity.getVelocity();
 
             float range = getSpell(Identifier.of(MOD_ID, "point_blank_shot")).range;
             int poscasterX = attacker.getBlockPos().getX();
@@ -31,16 +27,17 @@ public class PointBlankShot extends StatusEffect {
             int posentityX = entity.getBlockPos().getX();
             int posentityZ = entity.getBlockPos().getZ();
             int distance_to_target = (posentityX-poscasterX) + ( poscasterZ - posentityZ);
-            float diff_calc = (range - distance_to_target)/50;
 
-            float speed_leaping = -0.5F - diff_calc;
-            float leaping_height = 0.5F;
+            float diff_calc = (range - distance_to_target)/5;
+            var level = EnchantmentHelper.getLevel(Enchantments.POWER, attacker.getMainHandStack());
+            float knockback = 0.1F + diff_calc + level;
 
-            entity.velocityDirty = true;
-            entity.velocityModified = true;
-            entity.addVelocity(rotationVector.x * 0.1 + (rotationVector.x * 2.5 - velocity.x) * speed_leaping,
-                 leaping_height, rotationVector.z * 0.1 + (rotationVector.z * 2.5 - velocity.z) * speed_leaping);
-
+            double d = attacker.getX() - entity.getX();
+            double e;
+            for(e = attacker.getZ() - entity.getZ(); d * d + e * e < 1.0E-4; e = (Math.random() - Math.random()) * 0.01) {
+                d = (Math.random() - Math.random()) * 0.01;
+            }
+            entity.takeKnockback(knockback, d, e);
         }
     }
 }